Modifications réseau #1
Correction des bugs liés à la non disponibilité du serveur, envoi des
notifications textuels liés au client au client et non plus au serveur,
modification du design de la page de connexion et du menu principal,
possibilité de centrer automatiquement du texte en fonction de la
position donnée. (Apparition des talents dans le commit précédent)

+++ DW/Skills.cs
@@ -28,7 +28,7 @@ namespace DW
                 cheating = 5;
         }
 
-        public bool tryAction(string par1skill,int par2amount)
+        public bool tryAction(string par1skill,float par2amount)
         {
             bool v = true;
             if (par1skill == "wizardry")
